+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Add more debug logs for SentryViewHierarchy (#4780)
 - Add `sample_rand` to baggage (#4751)
+- Add `waitForFullDisplay` to `sentryTrace` view modifier (#4797)
 
 ### Fixes